+++ kernel/syscall.c
@@ -22,7 +22,7 @@
 ssize_t sys_user_print(const char* buf, size_t n) {
   // buf is now an address in user space of the given app's user stack,
   // so we have to transfer it into phisical address (kernel is running in direct mapping).
-  int hartid = read_tp();
+  uint64 hartid = read_tp();
   assert( current[hartid] );
   char* pa = (char*)user_va_to_pa((pagetable_t)(current[hartid]->pagetable), (void*)buf);
   sprint(pa);
@@ -35,7 +35,7 @@ ssize_t sys_user_print(const char* buf, size_t n) {
 static int ExitCount = 0;
 
 ssize_t sys_user_exit(uint64 code) {
-  int hartid = read_tp();
+  uint64 hartid = read_tp();
   sprint("hartid = %d: User exit with code:%d.\n", read_tp(), code);
   // in lab1, PKE considers only one app (one process). 
   // therefore, shutdown the system when the app calls exit()
@@ -51,7 +51,7 @@ ssize_t sys_user_exit(uint64 code) {
 // maybe, the simplest implementation of malloc in the world ... added @lab2_2
 //
 uint64 sys_user_allocate_page() {
-  int hartid = read_tp();
+  uint64 hartid = read_tp();
   void* pa = alloc_page();
   uint64 va = g_ufree_page;
   g_ufree_page += PGSIZE;
@@ -65,7 +65,7 @@ uint64 sys_user_allocate_page() {
 // reclaim a page, indicated by "va". added @lab2_2
 //
 uint64 sys_user_free_page(uint64 va) {
-  int hartid = read_tp();
+  uint64 hartid = read_tp();
   user_vm_unmap((pagetable_t)current[hartid]->pagetable, va, PGSIZE, 1);
   return 0;
 }
